refactor(view-menu): extract View menu into its own file, add divider above zoom
View menu construction moves out of main.swift's setupMenuBar into
ViewMenu.build() (Sources/edmd/App/ViewMenu.swift), mirroring the
existing FormatMenu.build() pattern. Also adds a separator between
Source Mode and Actual Size, so the zoom trio reads as its own group.
