+++ Sources/edmd/App/main.swift
@@ -269,46 +269,8 @@ class AppDelegate: NSObject, NSApplicationDelegate, NSMenuItemValidation {
         // Format menu — built from the declarative command registry.
         mainMenu.addItem(FormatMenu.build())
 
-        // View menu
-        let viewMenuItem = NSMenuItem()
-        let viewMenu = NSMenu(title: "View")
-
-        // Routes through the responder chain to the key window's toolbar.
-        viewMenu.addItem(withTitle: "Customize Toolbar…",
-                         action: #selector(NSWindow.runToolbarCustomizationPalette(_:)),
-                         keyEquivalent: "")
-        viewMenu.addItem(.separator())
-
-        let typewriterItem = viewMenu.addItem(
-            withTitle: "Typewriter Scroll",
-            action: #selector(AppDelegate.toggleTypewriterMode(_:)),
-            keyEquivalent: "")
-        typewriterItem.state = AppDelegate.typewriterModeEnabled() ? .on : .off
-
-        // View-mode toggle (Edit ↔ Read) + the Source-mode checkbox.
-        viewMenu.addItem(.separator())
-        viewMenu.addItem(FormatMenu.viewModeToggleItem())
-        viewMenu.addItem(withTitle: "Source Mode",
-                         action: #selector(Document.toggleSourceMode(_:)),
-                         keyEquivalent: "")
-
-        // Zoom (font size + max content width, scaled together). Target nil
-        // routes through the responder chain to the key window's Document.
-        // Kept last, directly above the separator AppKit inserts before its
-        // automatic "Enter/Exit Full Screen" item at the menu's end.
-        viewMenu.addItem(withTitle: "Actual Size",
-                         action: #selector(Document.actualSize(_:)),
-                         keyEquivalent: "0")
-        viewMenu.addItem(withTitle: "Zoom In",
-                         action: #selector(Document.zoomIn(_:)),
-                         keyEquivalent: "=")
-        viewMenu.addItem(withTitle: "Zoom Out",
-                         action: #selector(Document.zoomOut(_:)),
-                         keyEquivalent: "-")
-        viewMenu.addItem(.separator())
-
-        viewMenuItem.submenu = viewMenu
-        mainMenu.addItem(viewMenuItem)
+        // View menu — built in its own file (ViewMenu.swift).
+        mainMenu.addItem(ViewMenu.build())
 
         NSApplication.shared.mainMenu = mainMenu
     }
